Agregada alerta cuando se intenta acceder a la funcionalidad de favoritos sin estar logado.

+++ src/app/detalle/detalle.component.ts
@@ -4,6 +4,7 @@ import { TerrazasService } from '../terrazas.service';
 import { ActivatedRoute } from '@angular/router';
 import { JwtHelperService } from '@auth0/angular-jwt';
 import { FavoritosService } from '../favoritos.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-detalle',
@@ -23,6 +24,7 @@ export class DetalleComponent implements OnInit {
   esFavorito: boolean;
   usuarioId: string;
   terrazaId: string;
+  token: any;
 
   constructor(
     private terrazasService: TerrazasService,
@@ -49,14 +51,16 @@ export class DetalleComponent implements OnInit {
       try {
         this.terraza = await this.terrazasService.getTerrazaById(parseInt(params.idTerraza), objLocalStorage);
         // console.log(this.terraza);
-        const token = sessionStorage.getItem('token');
-        const jwt = new JwtHelperService();
-        const decodedToken = jwt.decodeToken(token);
-        this.usuarioId = decodedToken.userId;
-        this.terrazaId = params.idTerraza;
-        const getFavs = await this.favoritosService.getAll(this.usuarioId, this.terrazaId);
-        this.esFavorito = getFavs['BOOLEAN'];
-        // console.log('this.esFavorito', this.esFavorito);
+        this.token = sessionStorage.getItem('token');
+        if (this.token) {
+          const jwt = new JwtHelperService();
+          const decodedToken = jwt.decodeToken(this.token);
+          this.usuarioId = decodedToken.userId;
+          this.terrazaId = params.idTerraza;
+          const getFavs = await this.favoritosService.getAll(this.usuarioId, this.terrazaId);
+          this.esFavorito = getFavs['BOOLEAN'];
+          // console.log('this.esFavorito', this.esFavorito);
+        }
       } catch (err) {
         console.log(err);
       }
@@ -64,19 +68,39 @@ export class DetalleComponent implements OnInit {
   }
 
   async onClickFav() {
-    if (this.esFavorito === false) {
-      const result = await this.favoritosService.create(this.usuarioId, this.terrazaId);
-      this.esFavorito = true;
-      // console.log('result create', result);
+    console.log('this.token', this.token);
+
+    if (this.token) {
+      if (this.esFavorito === false) {
+        const result = await this.favoritosService.create(this.usuarioId, this.terrazaId);
+        this.esFavorito = true;
+        // console.log('result create', result);
+      } else {
+        const result = await this.favoritosService.delete(this.usuarioId, this.terrazaId);
+        this.esFavorito = false;
+        // console.log('result delete', result);
+      }
     } else {
-      const result = await this.favoritosService.delete(this.usuarioId, this.terrazaId);
-      this.esFavorito = false;
-      // console.log('result delete', result);
+      this.notLogged('Es necesario estar logado para agregar favoritos.');
     }
+  }
 
-
-
-
+  notLogged(msg) {
+    const Toast = Swal.mixin({
+      toast: true,
+      position: 'center',
+      showConfirmButton: false,
+      timer: 3000,
+      timerProgressBar: true,
+      onOpen: (toast) => {
+        toast.addEventListener('mouseenter', Swal.stopTimer)
+        toast.addEventListener('mouseleave', Swal.resumeTimer)
+      }
+    })
+    Toast.fire({
+      icon: 'error',
+      title: `<h4>${msg}</h4>`
+    });
   }
 
 }
